Fix Declare serialVersionUID in Serializable Class
In file: BaseEntity.java, the class BaseEntity is serializable but it does not contain any {serialVersionUID} field. The compiler generates one by default in such scenarios, but the generated id is dependent on compiler implementation and may cause unwanted problems during deserialization. iCR added the {serialVersionUID} field in the proper way. To learn more check the specification for Serializable.